Compute the ensemble mean and variance

This function is superseded by ens_stats. However, ens_mean_and_var() is still useful for computing the ensemble spread with a dropped member and is used by ens_spread_and_skill.

Usage

ens_mean_and_var(
  .fcst,
  mean_name = "ens_mean",
  var_name = "ens_var",
  sd_name = "ens_spread",
  var_drop_member = NULL
)

Arguments

.fcst

A harp_list object, or a harp_ens_point_df or harp_ens_grid_df data frame.

mean_name

The output column name for the ensemble mean

var_name

The output column name for the ensemble variance

sd_name

The output column name for the ensemble spread (standard deviation)

var_drop_member

Which members to drop for the calculation of the ensemble variance and standard deviation. For harp_fcst objects, this can be a numeric scalar - in which case it is recycled for all forecast models; a list or numeric vector of the same length as the harp_fcst object, or a named list with the names corresponding to names in the harp_fcst object.

...

Value

A harp_fcst object with columns ens_mean and ens_var added to the forecast tables.

Details

The ensemble mean and variance are computed and added as columns to tables in a harp_df object or harp_list object..