Mutating joins for harp_lists

These are the harp_list methods for join functions, for example inner_join from the dplyr package. In the case of x being a harp_list, y can either be a data frame or another harp_list that is the same length of x. If y is a data frame, an attempt will be made to join y to each data frame in the harp_list, x. If y is a harp_list, data frames in corresponding elements of x and y will be joined. Note that this is done on the basis of location in the list and names are not taken into account. Names in the output are always taken from the names of x.

Usage

# S3 method for harp_list
inner_join(x, y, by = NULL, ...)

# S3 method for harp_list
left_join(x, y, by = NULL, ...)

# S3 method for harp_list
right_join(x, y, by = NULL, ...)

# S3 method for harp_list
full_join(x, y, by = NULL, ...)

Arguments

x, y

A pair of data frames, data frame extensions (e.g. a tibble), or lazy data frames (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

by

A join specification created with join_by(), or a character vector of variables to join by.

If NULL, the default, *_join() will perform a natural join, using all variables in common across x and y. A message lists the variables so that you can check they're correct; suppress the message by supplying by explicitly.

To join on different variables between x and y, use a join_by() specification. For example, join_by(a == b) will match x$a to y$b.

To join by multiple variables, use a join_by() specification with multiple expressions. For example, join_by(a == b, c == d) will match x$a to y$b and x$c to y$d. If the column names are the same between x and y, you can shorten this by listing only the variable names, like join_by(a, c).

join_by() can also be used to perform inequality, rolling, and overlap joins. See the documentation at ?join_by for details on these types of joins.

For simple equality joins, you can alternatively specify a character vector of variable names to join by. For example, by = c("a", "b") joins x$a to y$a and x$b to y$b. If variable names differ between x and y, use a named character vector like by = c("x_a" = "y_a", "x_b" = "y_b").

To perform a cross-join, generating all combinations of x and y, see cross_join().

...

Arguments passed on to dplyr::inner_join, dplyr::left_join, dplyr::right_join, dplyr::full_join

suffix

If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

keep

Should the join keys from both x and y be preserved in the output?

If NULL, the default, joins on equality retain only the keys from x, while joins on inequality retain the keys from both inputs.

If TRUE, all keys from both inputs are retained.

If FALSE, only keys from x are retained. For right and full joins, the data in key columns corresponding to rows that only exist in y are merged into the key columns from x. Can't be used when joining on inequality conditions.

Value

A harp_list with the same names as x.